bounds.c: move casts from pointers to ints to pointer-larger-int directory (disabled for Win32)
