+++ tests/typechecking/pointer-larger-int/lit.local.cfg
@@ -1,0 +1,8 @@
+import re
+
+# On 32-bit windows a pointer is not larger than int
+if not re.match(r'^i686-pc.*-(windows-msvc|windows-gnu)$', config.target_triple):
+    config.available_features.add('pointer-larger-int')
+
+if not 'pointer-larger-int' in config.available_features:
+    config.unsupported = True
